Added spinner for selecting projects in the editor activity.
We can use the stock ArrayAdapter, because the Project.toString() method returns all we need to see.

Some extra work was required to pre-select the correct project in the spinner. It seems that the project description is not always constant throughout the application. When calling getWeekOverview, the list of all available projects contains different descriptions than those that are in the list if TimeSheetRows. For example: 'Internal project' versus 'Internal Projects'.