Fixed issue with importing the project in Android Studio.
The build.gradle was using the System.console() to read the passwords for signing releases. This caused a NullPointerException in Android Studio, making it impossible to import the project.

The solution is lifted from https://www.timroes.de/2013/09/22/handling-signing-configs-with-gradle/. The signingConfig information is moved to misc/signing/signing.gradle file, which is not version controlled. The build.gradle has been modified so that the release signing config is only set up when the signing.gradle file is present on the system. On other systems, no errors are thrown anymore.