Added option for deleting time sheet entries in the editor action bar.
The menu option kicks off the DeleteEntryTask to perform the delete request. The option is only visible when not creating a new entry. After the request completes, the DayOverviewActivity gets a special result code to indicate the entry was deleted.

To keep the EditTimeSheetFragment as compact as possible, the SaveEntryTask was also moved out of the EditTimeSheetFragment.

Fixes issue #3.