Created overview activity for week and month overviews
The WeekOverviewActivity was discarded in favor of a general overview activity. The user can switch between the default week overview fragment and a new month overview fragment.

A workaround was required to make sure the viewpager for the week overview fragment in this new setting. The FragmentWithChildFragmentManager resets the ChildFragmentManager whenever the fragment is attached to a layout. Without this workaround, the viewpager would not refresh when the user navigated to the week overview fragment.