Supermassive refactoring of data model.
Inverted the data model so the TimeEntry object is the center piece. TimeSheetRow is refurbished into Task, and no longer contains a reference to the TimeEntry.

This complicates parsing and storing the data from XTime backend a bit, but makes more sense for the actual use of the data. It is now possible to query the TimeEntries and get useful data immediately.